pyproject.toml
[project]
name = "gsuite-mcp"
version = "0.1.0"
description = "A comprehensive MCP server for Google Workspace APIs (Gmail, Calendar, People)"
readme = "README.md"
authors = [
    { name = "Harper Reed", email = "harper@nata2.org" }
]
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.23.1",
]

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
